Use Thor for built-in Time zone tasks
Currently we use both Thor and Rake for `bin/rails` commands.
We eventually want to get all the built-ins task promoted to Thor Commands.
This migrates the `time:zones` tasks to Thor.

This also fixes passing `US` or `-8` as an argument, as the Rake
implementation was broken. The arguments have to be passed without
square brackets though.

As there are a lot of subcommands and arguments a USAGE file was added
to show more example usages.

Co-authored-by: Jonathan Hefner <jonathan@hefner.pro>