Converted StringInquirer into a PORO with better performance.
Currently, ActiveSupport::StringInquirer inherit from String, and use
method_missing on all of its calls, so we can have a performance
problem.

Converting it to a PORO and creating a method on the instance with the
name of the positive result and leaving just a test to return false to
predicate methods can speed up operations.

The current benchmark result returned for the new benchmark test is:

all true  0.046796   0.000000   0.046796 (  0.047182)
all false 0.045121   0.000017   0.045138 (  0.045299)
half      0.042590   0.000000   0.042590 (  0.042956)

with this change it results in:

all true  0.013614   0.000000   0.013614 (  0.013640)
all false 0.029838   0.000008   0.029846 (  0.030124)
half      0.017558   0.000000   0.017558 (  0.017680)