Fix rails-ujs tests not being linted
Before rails-ujs was [updated][1] from coffeescript to es6, the lint
task ran eslint on both the coffee files and the ujs test files. When it
was updated to lint the new es6 files, the test files were erroneously
left out.

This commit fixes the task to lint the rails-ujs test files and fixes
the lint errors that have been introduced since.

[1]: 7d116c93cf6cf2470600860c4c17417df7768c34