Add `test:prepare` to `bin/setup` template
Suggested by @etiennebarrie [here](https://github.com/rails/rails/pull/47210#issuecomment-1476420481), I think this is a good idea. This way if you clone a repo and want to run a single test, you will be able to successfully do so if you ran `bin/setup` first (which is a good idea as it also tries to set up everything else you need).