Generalize check for foreign keys wgeneralize check for foreign keys
Fixes https://github.com/rails/rails/issues/47574

It was reported that when you have an association that uses a scope, querying for inverses do not correctly return the
same object.

@andrewberls provided the reproduction steps in the issue that was filed.

It turns out, when rails needs to check if the foreign key matches the record in `matches_foreign_key?` it uses the `id`
of the record. This works majority of the time but unfortunately does not work when the foreign key is not an `id` but
something else like a `token` or a `uuid`.

This change updates the check to call `association_primary_key` from the reflection to properly match the record.

Co-authored-by: Adam Hess <adamhess1991@gmail.com>