Refactor Relation#none to no longer use Object#extend
Ref: https://github.com/rails/rails/pull/47314
Ref: https://bugs.ruby-lang.org/issues/19436

`Object#extend` is really bad for performance as it generate an
infinite amount of unique classes from the point of view of the Ruby
virtual machine, so it keeps invalidating call caches.

Additionally, these call caches keep a reference on the last instance
they saw, which may lead to what is essentially a memory leak.