Generate unique members of a composite key in fixtures
Typically, a fixture's label is hashed into a single unique identifier
that sufficiently represent the respective fixture. In composite key
models, there's still only one label, but a requirement to populate
values for an arbitrary number of columns. Until now, these columns
would all take on the same value, but these may appear odd to an end
user and lead to confusion. This commit uses the index of the column in
the composite key to shift the hash in a deterministic way. In this
case, we bit shift the hashed value index-times and modulo it with the
max. This differentiates components of the key enough that there doesn't
appear to be a correlation.