Fix autoincrement on primary key
MySQL adapters take an auto_increment option but when using a primary
key type that option is silently ignored and doesn't change behavior.
This fixes the issue by only applying `options[:auto_increment] = true`
if it wasn't set to false already.

I didn't make changes to the other adapters because they don't accept
`auto_increment` as an option. If we want this for postgres we'll need
to implement the option first.