Only default to select primary key when non-composite
When the relation being handled has no selected values, it fallsback to selecting the model's primary key. In composite key cases, this can lead to attempting to compare a single attribute to a list of attributes, which leads to malformed syntax at the SQL level. This check prevents any code from attempting to do so, by raising in CPK cases when defaulting to primary key. For now, users can build these queries themselves.