Add `COPY .. TO` to read queries for PostgreSQL
Fixes https://github.com/rails/rails/issues/47870

It was pointed out in the report that

> COPY moves data between PostgreSQL tables and standard file-system files. COPY TO copies the contents of a table to a file, while COPY FROM copies data from a file to a table (appending the data to whatever is in the table already). COPY TO can also copy the results of a SELECT query.

Using COPY .. TO with a read only connection should not be a problem because we do not write to the database but only
read the records that will be written somewhere else.

This change adds the regex `/copy.*to/i` to the list of read only markers for postgres. I decided to use a regex here so
that we can also make sure that `COPY .. FROM` cannot be used with a read only connection.