Add support in Arel for materialized subqueries in CTEs
Postgres and SQLite both support a non-standard extension to the
CTE syntax to indicate that a CTE subquery should be materialized,
i.e., not folded into the main query but evaluated separately.
This can be useful in cases where the query planner would
otherwise make poor decisions.

The syntax, in both databases, is:
  `WITH foo AS MATERIALIZED (...) ...`

This PR adds support by introducing a new unary Materialized node
that can wrap a subquery like so:

   ```ruby
   posts = Arel::Table.new(:posts)
   comments = Arel::Table.new(:comments)
   good_comments = Arel::Table.new(:good_comments)
   subquery = comments.project(Arel.star).where(comments[:rating].gt(7))
   materialized_subquery = Arel::Nodes::Materialized.new(subquery)

   posts.
     project(Arel.star).
     with(Arel::Nodes::As.new(good_comments, materialized_subquery)).
     where(posts[:id].in(good_comments.project(:post_id))).
     to_sql

   # "WITH \"good_comments\" AS MATERIALIZED (SELECT * FROM \"comments\" WHERE \"comments\".\"rating\" > 7) SELECT * FROM \"posts\" WHERE \"posts\".\"id\" IN (SELECT post_id FROM \"good_comments\")"
   ```