Fix backport of 47948 including other change
The backports of a [bugfix][1] for migrations to [7-0-stable][2] and
[6-1-stable][3] ended up including another [change][4] from main, which
is now causing CI to fail for those stable branches.

This commit fixes the CI issues by un-including the additional change
from main.

An alternative to this fix would be to backport the test changes from
the additional commit, but since it changes error messages it seems
better to leave that on the main branch.

[1]: 11aaa3db6ca350fe7bd0be93ce2a9e8a9c17ab61
[2]: e2a99b56e57a0517d88fc0ff9bc111704e44d7df
[3]: 33a9277546644bd2fac328f1c764a3323bad3eea
[4]: 39a20095bf006de816eec32b4d431226b80f10cf