Better deprecation warning for `ActiveSupport::Deprecation.warn`
ref: https://github.com/rails/rails/pull/47354#discussion_r1170553124

It is common to see this code in apps (not in gems/engines, but in user code):

```ruby
ActiveSupport::Deprecation.warn "foo"
```

Currently that code will emit this warning:

> DEPRECATION WARNING: Calling warn on ActiveSupport::Deprecation is deprecated and will be removed from Rails (use your own Deprecation object instead)

This PR changes the warning so that it gives you more direct advice on how to fix it:

> DEPRECATION WARNING: Calling warn on ActiveSupport::Deprecation is deprecated and will be removed from Rails (use `ActiveSupport.deprecator.warn` instead)

My understanding is that this is fine for user code, and we are not expecting users to define their own deprecators for in-app use. Given that, we should just tell the user what to do in the error message.