Active storage proxy setting filename
The current version of the guides states that we can set proxying for
ActiveStorage. However, it mentions that this is set in
config/initializers/active_storage.rb

This file does not seem to exist, so I suggest that this config is set
in one of the environment files. In the case of proxying it makes sense
to set this in config/environments/production.rb.

Alternatively it could be set in config/application.rb