Call proxy attribute methods for reserved aliases
This commit allows aliasing columns named with one of the reserved
methods like `id`, `id_before_type_cast`, `id_was`, etc if database
schema happed to have one of these.

Currently there is no convenient way of reading the `id` column value
which is not being used as a primary key as `id` in Rails is a reserved
term which means "an identifier" or the value stored in the primary key
column.

Given a `posts` table with both `id` and `title` columns where `title`
is being used as a primary key it is possible to alias `id` attribute
to a new name so we can fetch the column value.

For example:
```ruby
class Post < ActiveRecord::Base
  self.primary_key = :title
  alias_attribute :post_id, :id
end

post = Post.create(title: "Hello world")
post.post_id # => 1
```

Co-authored-by: Paarth Madan <paarth.madan@shopify.com>