Improve SauceLabs connection handling for CI
When debugging the SauceLabs integration in my local CI environment, I discovered this option for connection logs.

Also, the example from the karma-sauce-launcher recommends adding this check for the environment variables:
https://github.com/karma-runner/karma-sauce-launcher/blob/69dcb822a45d29e57297b0eda7af4123ae55aafd/examples/karma.conf-ci.js#L2-L5

Even though we hard code the SauceLabs connection info for CI, that doesn't guarantee it will always be there in the same place, or work.