Fix `Marshal.dump` on records that use AR Encryption
Fixes https://github.com/rails/rails/issues/48038

I don't think we need to use a proc here - it doesn't seem like there is any performance benefit in doing so. The default is being read from the `columns_hash`, which will already be loaded in memory. And then the `default` method is just reading a scalar value (or returning nil).