Bump karma and dependencies
Changes to the config were required to get latest firefox running:
https://saucelabs.com/resources/blog/update-firefox-tests-before-oct-4-2022

Thanks to hotwired/stimulus#601 and trix I was able to get the config working, the secret was that resolution for `webdriverio` needs to go in the root `package.json`.

In the short-to-mid term I'd like to explore updating our builds to use the latest yarn, or remove it entirely, as that seems to be a common thread in these issues.

There was another investigation into mitigating the flaky saucelabs failures in #51990 but I think let's make sure we're on the latest version and browsers to see if that improves things or provides new and interesting error states first.