Raise retryable exception class on abrupt Postgres connection closure
👋 Hi team! Fresh off RailsConf here. Enjoyed some of the convos
with matthewd and figured I'd check out the new connection and query retry feature
on rails@main. We have monkey patched ActiveRecord internally to perform query retry and
connection retry to recover from failures automatically with Postgres.

While testing, I came across the following issue:

Postgres can at times throw the following error.
```
server closed the connection unexpectedly
        This probably means the server terminated abnormally
        before or while processing the request.
```

This can happen during a database crash or a failover, or when connections
are intentionally terminated. When this happens, an already checked out ActiveRecord
would raise `ActiveRecord::StatementInvalid: PG::UnableToSend: server closed the connection unexpectedly`.

Raising an `ActiveRecord::StatementInvalid` means that the query will not get retried,
even when `allow_retry` is set to `true` because only `ConnectionNotEstablished` or `ConnectionFailed`
are considered for retries (via `retryable_connection_error?`).

Hence, with this change, we will now also throw a `ConnectionNotEstablished` by extending
the already existing error message matching for when Postgres is down.

In this PR, we also use `pg_terminate_backend` to recreate the a database failover
scenario in the test. Which would prompt `pg` gem to throw `PG::UnableToSend: server closed the connection unexpectedly`.
Accordingly we rest and retry and recover from the queries.