More Subscriber and LogSubscriber doc uniformity
attach_to was previously [improved][1] to allow defining methods after
attaching instead of having to attach after all methods have been
defined. The docs for Subscriber were [updated][2], but LogSubscriber
docs were not.

This commit copies the attach_to doc changes from Subscriber to
LogSubscriber. In addition, there are other improvements to linking and
applying the monospace formatting that were present in one of Subscriber
or LogSubscriber and are now applied to both. The final change is
updating the description of how flush_all! is used, because its usage
has [changed][3] since this doc was written.

[1]: 34088572270a1dd5a2164b6aa5fc3642cb0479cb
[2]: 25b3f738e4f5b09e4d6a66e1454e697defcdda2c
[3]: 378464a2e47bb849f3351cb8c87366554b7ce74d