Discard connection when in an unknown transaction state.
Fixes #48164

Introduces the concept of a transaction lock. This is used to detect
cases where `within_new_transaction` raises, potentially leaving the
connection in an untracked, open transaction.

The idea is to store and acquire a new lock before attempting to
materialize a transaction, and to only discard the lock once the
transaction state is definitely known. If `within_new_transaction`
raises, without the lock being discarded, subsequent attempts to
create a transaction will detect this case and discard the connection.