Optimize mime types validation in ActionView::LookupContext
Fix: https://github.com/rails/rails/issues/48156

The assumption here is that in the overwhelming majority of
cases, all formats are valid.

So we first check if any of the formats is invalid before duping
the details hash and filtering them.

Additonally, by exposing a (non-public) `valid_symbols?` method, we
can check symbols are valid without resporting to `Array#%` which
would needlessly duplicate the `formats` array.